+++ src/utils.js
@@ -8,7 +8,7 @@ import { saveAs } from 'file-saver';
 const DEFAULT_OPTIONS = {
   filename: 'Image',
   forceFixText: false,
-  printDate: true,
+  printDate: true
 };
 
 /**
@@ -59,7 +59,7 @@ const hardFixText = (
     'mark',
     'p',
     'small',
-    'strong',
+    'strong'
   ]
 ) => {
   const allTexts = node.querySelectorAll(seek);
@@ -141,7 +141,7 @@ const applyFixs = (node, forceFixText = false) => {
 const getOptions = (userOptions) => {
   return {
     ...DEFAULT_OPTIONS,
-    ...userOptions,
+    ...userOptions
   };
 };